+++ scvelo/datasets/__init__.py
@@ -1,0 +1,17 @@
+from ._datasets import (
+    dentategyrus,
+    forebrain,
+    pancreas,
+    pancreatic_endocrinogenesis,
+    toy_data,
+)
+from ._simulate import simulation
+
+__all__ = [
+    "dentategyrus",
+    "forebrain",
+    "pancreas",
+    "pancreatic_endocrinogenesis",
+    "simulation",
+    "toy_data",
+]
